Advancing on background renderer.
Surprisingly hard to extend the API, and a lot of work for common
functionality.
This should be extracted to the warhol API after this.

+++ warhol/graphics/common/render_command.h
@@ -66,8 +66,8 @@ struct MeshRenderAction {
   IndexRange index_range;
 
   // The counts of this are defined in the corresponding shader.
-  float* vert_uniforms = nullptr;
-  float* frag_uniforms = nullptr;
+  uint8_t* vert_uniforms = nullptr;
+  uint8_t* frag_uniforms = nullptr;
   Texture* textures = nullptr;
 };
